requirements.txt
pydantic[email]
sqlmodel
sqlmodel-controller
sqlalchemy
bcrypt
PyJWT
streamlit
requests
pytest
pytest-cov